Wild Oasis boiler plate
I have initialized a new Vite application for my wild oasis application that will be a hotel booking site in essence. I have created what i think will be the majority of the necessary components, folders and helper files. I plan to learn supabase for this application and implement many of the features i have been learning so far.