Layout + React router
I have added a few more components to be reused throughout the application such as the Input and Button component. I have also added an applayout page to contain my routes and control the layout and flow of the application.

I have also added a GlobalStyles component to contain my global reusable styles.