Main navigation
I have added the main navbar that the users will use to get around the site which uses react router and without updating the entire page each time can navigate the user accordingly placing the content in the Outlet component. I have also implemented a logo and renamed the site to wild oasis. Next up will be setting up and structuring the back end, for which i will be learning and using Supabase.