React Query integration.
An issue with my last few major projects was remote state management, often causing optimization issues when loading data for multiple components.

Thankfully now i have become aware of a way to cache this data so i only have to fetch it once and when the data is changed in some way. Which i feel will help a lot with this issue.

To do this i have gone with the react-query library to manage the remote state we are getting from supabase.

I really cant take much credit here because supabase's API docs make most of this process fairly trivial as it automatically gives you the necessary query syntax to be copy and pasted.

Then i use react query to fetch and mutate the data using the functions supabase gives us. So far i have implemented the fetching and deleting of cabins for testing. It all works great so i will carry on.

I also added the cabins page's table that will eventually show all the cabins on offer and put a few test rows in for now, which will be replaced with the real data later.

On the supabase end for now i have allowed any user to delete but later this will be restricted only to people with the right authorization.