Adding cabins
I have added the ability for staff members to add cabins as well as delete them. To do this i have used a couple packages, the first is react-hot-toast to handle the user notifications. Then i have added react-hook-form to manage the creating and submitting of the cabin form as well as to handle the errors with ease.

I have validated all input fields, and displayed the necessary error messages if there is an issue.

This is now all linked to supabase through my apiCabins file, which is responsible for inserting the new row into the supabase db.