Image uploading and cabin editing
I have added the image uploading feature using supabase' storage buckets. Again supabase API docs make this all really simple so i cant take much credit.

I have also added the ability for editing of cabins. And made it work with the images such that the image is not required for an edit and if no changes are made the image path it will simply use the image already stored for that cabin.