Custom hooks and Settings.
I have abstracted some of the database communication logic to their own custom hooks. This really cleans up the component file and makes it much easier to access the data such as the cabin data anywhere in the application.

I have also added the settings page and allowed the staff members to edit the settings such as the breakfast price etc.

I also added the ability to duplicate a cabin to make it much easier to add new ones by simply duplicating and changing the necessary information.