Advanced React Patterns Modal
Recently i have been delving into some advanced react patterns. i have decided to employ one of them in this project to achieve a reusable modal window that is a compound component.

Firstly i cleaned up the AddCabin component to ensure its clean and no longer handling any logic.

I then moved this logic into the modal component. When making a compound component there are 3 steps in my case 4 because of how i have linked the components which is to make a context where my state will live. Then i make the parent component in this case the Modal and then we make the child helper components. In this case the Open and Window components. Finally i link them together. Using this i can now easily reuse this modal across the application. To further aid that i have used a portal for the Window. This will keep the component in the same place in the react tree but not the DOM keeping the styles and state consistent wherever it is used.