Custom hook for closing modal
I wanted the modal to be able to close when the user clicks outside of the modal as well as the close/cancel buttons. To do so i have made a custom hook which is responsible for that by having a useEffect listen for clicks where it is not equal to the modal component a.k.a the ref i have made. If that is true and the ref exists then it will close the modal.