Context menu + Reusability changes
I have had less free time to code of late sadly so i have been trying to get as much as i can done when i can. Today i have:

Added a context menu in place of the 3 buttons that were previously clogging up each cabin row. It is now a clean 3 dots that will open the menu offering the same 3 options as before to duplicated edit or delete. To achieve this i have used a compound menus component that contains all the logic for the buttons, toggling the menu etc. To get the position of where each menu should be placed i have use the getBoundingClientRect together with some hacky math.

I have made use of my new reusable modal window to also implement the confitm delete component which now pops up as a modal to confirm with the user that they do indeed want to proceed with a harmful action. This comp is also very reusable for later when confirming similar things.

I have also refactored the Table component to also be much more reusable as i will likely want to reuse something for similar to this for the bookings part of the application. To make this work in a reusable way i  used the render props pattern to tell the Table.Body component how it should render the children.