Filtering and Sorting
Added functionality for filtering the cabins. I tried to make this very re-usable both for use later in this project and with another that i will soon be starting. For this i have made a filter styled component which takes in a field name and an array of option objects from which it will render the buttons for the filtering i.e. filtering by discounts. I am accessing which option was chosen by using the useSearchParams hook to get the necessary info from the URL.

To couple with the sorting feature i also added a means of sorting the cabins based on a few different things much the same way as the filtering to be reusable.