Server side filtering and sorting
Today i have implemented the bookings table and used a component for development to upload some dummy table data for me to work with.

I have also implemented the filter and sort components that were used on the cabin table here but re-worked them to do the filtering on the server side as here we dont want to get more data from the database than we need to improve optimization. Thankfully supabase is incredibly intuitive so this is no problem, we simply tailor the query string as needed and let react query re-fetch the necessary data. I ran into an issue where it was not triggering a re-render upon the data change. I figured out that query gives me a solution for this as well to give it a dependency array essentially, telling it to refetch and thus re-render the page whenever the filter or sortBy fields change.