Bookings Pagination
I have used a combination of react query and supabase functionality to get the count for the results as well as paginate them on the api side and not the front end. This also allows me to prefetch the pages to reduce load times when moving between pages.

I have added a styled pagination component for the user to see the page number and number of results as well as navigate through the pages at will.